small fixes for Windows 64-bit compilation
git-svn-id: http://dev.corsaair.com/svn/redtamarin/trunk@1644 8266667d-5eeb-4550-8638-74b212e0aa7d

+++ src/shell/ProgramClass.cpp
@@ -123,12 +123,12 @@ namespace avmshell
         Stringp output = core()->newStringUTF8( "" );
         
         if (read_fp != NULL) {
-            chars_read = fread(buffer, sizeof(char), BUFSIZ, read_fp);
+            chars_read = (int) fread(buffer, sizeof(char), BUFSIZ, read_fp);
             output = output->append( core()->newStringUTF8(buffer, chars_read) );
             
             while(chars_read > 0) {
                 buffer[chars_read - 1] = '\0';
-                chars_read = fread(buffer, sizeof(char), BUFSIZ, read_fp);
+                chars_read = (int) fread(buffer, sizeof(char), BUFSIZ, read_fp);
                 output = output->append( core()->newStringUTF8(buffer, chars_read) );
             }
